+++ src/useOnAll/index.ts
@@ -2,17 +2,20 @@ import { Collection } from 'typesaurus/collection'
 import { Doc } from 'typesaurus/doc'
 import onAll from 'typesaurus/onAll'
 import { useEffect, useState } from '../adaptor'
+import { TypesaurusHookResult } from '../types'
 
 export default function useOnAll<Model>(
   collection: Collection<Model>
-): Doc<Model>[] | undefined {
+): TypesaurusHookResult<Doc<Model>[] | undefined> {
   const [result, setResult] = useState<Doc<Model>[] | undefined>(undefined)
+  const [error, setError] = useState<unknown>(undefined)
+  const loading = result === undefined
 
   const deps = [JSON.stringify(collection)]
   useEffect(() => {
     if (result) setResult(undefined)
-    return onAll(collection, setResult)
+    return onAll(collection, setResult, setError)
   }, deps)
 
-  return result
+  return [result, { loading, error }]
 }