+++ test/setupJestLocal.ts
@@ -1,0 +1,10 @@
+import { injectTestingAdaptor } from 'typesaurus/testing'
+import * as testing from '@firebase/testing'
+
+injectTestingAdaptor(testing.initializeAdminApp({ projectId: 'project-id' }))
+injectTestingAdaptor(
+  testing.initializeTestApp({
+    projectId: 'project-id',
+    auth: { uid: 'user-id' }
+  })
+)